+++ pricing/src/epilot/models/shared/salestax.py
@@ -4,11 +4,6 @@ from __future__ import annotations
 from enum import Enum
 
 class SalesTax(str, Enum):
-    r"""The default tax rate applicable to the product.
-    This field is deprecated, use the new `tax` attribute.
-
-    Deprecated class: This will be removed in a future release, please migrate away from it as soon as possible.
-    """
     NONTAXABLE = 'nontaxable'
     REDUCED = 'reduced'
     STANDARD = 'standard'